# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MonkeyCode API",
        "title": "MonkeyCode API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/model": {
            "get": {
                "description": "获取模型列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取模型列表",
                "operationId": "list-model",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.AllModelResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "更新模型",
                "operationId": "update-model",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "创建模型",
                "operationId": "create-model",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/model/check": {
            "post": {
                "description": "检查模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "检查模型",
                "operationId": "check-model",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CheckModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Model"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/model/my": {
            "get": {
                "description": "获取我的模型列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取我的模型列表",
                "operationId": "my-model-list",
                "parameters": [
                    {
                        "enum": [
                            "llm",
                            "coder",
                            "embedding",
                            "audio",
                            "reranker"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "ModelTypeLLM",
                            "ModelTypeCoder",
                            "ModelTypeEmbedding",
                            "ModelTypeAudio",
                            "ModelTypeReranker"
                        ],
                        "description": "模型类型 llm:对话模型 coder:代码模型",
                        "name": "model_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Model"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/model/token-usage": {
            "get": {
                "description": "获取模型token使用情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取模型token使用情况",
                "operationId": "get-token-usage",
                "parameters": [
                    {
                        "enum": [
                            "llm",
                            "coder",
                            "embedding",
                            "audio",
                            "reranker"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "ModelTypeLLM",
                            "ModelTypeCoder",
                            "ModelTypeEmbedding",
                            "ModelTypeAudio",
                            "ModelTypeReranker"
                        ],
                        "description": "模型类型 llm:对话模型 coder:代码模型",
                        "name": "model_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ModelTokenUsageResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/completion/accept": {
            "post": {
                "description": "接受补全请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAIV1"
                ],
                "summary": "接受补全请求",
                "operationId": "accept-completion",
                "parameters": [
                    {
                        "description": "补全请求",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AcceptCompletionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "consts.ModelStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive"
            ],
            "x-enum-varnames": [
                "ModelStatusActive",
                "ModelStatusInactive"
            ]
        },
        "consts.ModelType": {
            "type": "string",
            "enum": [
                "llm",
                "coder",
                "embedding",
                "audio",
                "reranker"
            ],
            "x-enum-varnames": [
                "ModelTypeLLM",
                "ModelTypeCoder",
                "ModelTypeEmbedding",
                "ModelTypeAudio",
                "ModelTypeReranker"
            ]
        },
...
        "domain.AcceptCompletionReq": {
            "type": "object",
            "properties": {
                "completion": {
                    "description": "补全内容",
                    "type": "string"
                },
                "id": {
                    "description": "记录ID",
                    "type": "string"
                }
            }
        },
...
        "domain.AllModelResp": {
            "type": "object",
            "properties": {
                "providers": {
                    "description": "提供商列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ProviderModel"
                    }
                }
            }
        },
...
        "domain.CheckModelReq": {
            "type": "object",
            "required": [
                "api_base",
                "api_key",
                "model_name",
                "provider"
            ],
            "properties": {
                "api_base": {
                    "description": "接口地址",
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥",
                    "type": "string"
                },
                "model_name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                }
            }
        },
...
        "domain.CreateModelReq": {
            "type": "object",
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "model_name": {
                    "description": "模型名称 如: deepseek-v3",
                    "type": "string"
                },
                "model_type": {
                    "description": "模型类型 llm:对话模型 coder:代码模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelType"
                        }
                    ]
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                }
            }
        },
...
        "domain.Model": {
            "type": "object",
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "id": {
                    "description": "模型ID",
                    "type": "string"
                },
                "input": {
                    "description": "输入token数",
                    "type": "integer"
                },
                "is_active": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "model_name": {
                    "description": "模型名称 如: deepseek-v3",
                    "type": "string"
                },
                "model_type": {
                    "description": "模型类型 llm:对话模型 coder:代码模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelType"
                        }
                    ]
                },
                "output": {
                    "description": "输出token数",
                    "type": "integer"
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                },
                "status": {
                    "description": "状态 active:启用 inactive:禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelStatus"
                        }
                    ]
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "integer"
                }
            }
        },
        "domain.ModelBasic": {
            "type": "object",
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                }
            }
        },
...
        "domain.ModelTokenUsage": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "description": "时间戳",
                    "type": "integer"
                },
                "tokens": {
                    "description": "使用token数",
                    "type": "integer"
                }
            }
        },
        "domain.ModelTokenUsageResp": {
            "type": "object",
            "properties": {
                "input_usage": {
                    "description": "输入token使用记录",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ModelTokenUsage"
                    }
                },
                "output_usage": {
                    "description": "输出token使用记录",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ModelTokenUsage"
                    }
                },
                "total_input": {
                    "description": "总输入token数",
                    "type": "integer"
                },
                "total_output": {
                    "description": "总输出token数",
                    "type": "integer"
                }
            }
        },
        "domain.ProviderModel": {
            "type": "object",
            "properties": {
                "models": {
                    "description": "模型列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ModelBasic"
                    }
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateModelReq": {
            "type": "object",
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "id": {
                    "description": "模型ID",
                    "type": "string"
                },
                "model_name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                },
                "status": {
                    "description": "状态 active:启用 inactive:禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelStatus"
                        }
                    ]
                }
            }
        },
...
        "web.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}